Complex class and the inherited Load class
this assignment defines a complex class with the attributes of a complex number (real and imaginary parts)
then an inherited class loads with capacitor, inductor and resistor describe and define the functionality of each electrical element as a comlex number consists of the resistance and the reactance.
the resistor has zero reactance
while the inductor has a positive reactance in a relationship between the angular frequency and the inductance
the capacitor has negative reactance with a relationship between the frequency and the capacitance.
these java files are used for electrical circuits analysis